Changed name of package, restructured library
