+++ src/common/angular_velocity_vector.rs
