+++ src/common/dis_error.rs
