+++ src/common/entity_coordinate_vector.rs
