+++ src/entity_information/entity_appearance.rs
