+++ src/entity_information/general_appearance.rs
