+++ src/entity_information/specific_appearance.rs
