+++ src/lib.rs
@@ -1,1 +1,14 @@
-pub mod dis;
+//     open-dis-rust - Rust implementation of the IEEE-1278.1 Distributed Interactive Simulation
+//                     (DIS) application protocol v6 and v7
+//     Copyright (C) 2023 Cameron Howell
+//
+//     Licensed under the BSD 2-Clause License
+
+pub mod common;
+pub mod distributed_emissions;
+pub mod entity_information;
+pub mod logistics;
+pub mod radio_communications;
+pub mod simulation_management;
+pub mod simulation_management_with_reliability;
+pub mod warfare;
