+++ src/simulation_management/action_request_pdu.rs
@@ -1,41 +1,57 @@
-use crate::dis::common::{
+use bytes::{Buf, BufMut, BytesMut};
+use std::any::Any;
+
+use crate::common::{
     dis_error::DISError,
     entity_id::EntityId,
     pdu::Pdu,
     pdu_header::{PduHeader, PduType, ProtocolFamily},
 };
-use bytes::{Buf, BufMut, BytesMut};
-use std::any::Any;
 
 #[derive(Copy, Clone, Debug)]
-pub struct CreateEntityPdu {
+pub struct ActionRequestPdu {
     pub pdu_header: PduHeader,
     pub originating_entity_id: EntityId,
     pub receiving_entity_id: EntityId,
     pub request_id: u32,
+    pub action_id: u32,
+    pub number_of_fixed_datum_records: u32,
+    pub number_of_variable_datum_records: u32,
+    pub fixed_datum_records: u64,
+    pub variable_datum_records: u64,
 }
 
-impl CreateEntityPdu {
+impl ActionRequestPdu {
     pub fn default() -> Self {
-        CreateEntityPdu {
+        ActionRequestPdu {
             pdu_header: PduHeader::default(
-                PduType::CreateEntity,
+                PduType::ActionRequest,
                 ProtocolFamily::SimulationManagement,
                 56,
             ),
             originating_entity_id: EntityId::default(1),
             receiving_entity_id: EntityId::default(2),
             request_id: 0,
+            action_id: 0,
+            number_of_fixed_datum_records: 0,
+            number_of_variable_datum_records: 0,
+            fixed_datum_records: 0,
+            variable_datum_records: 0,
         }
     }
 }
 
-impl Pdu for CreateEntityPdu {
+impl Pdu for ActionRequestPdu {
     fn serialize(&self, buf: &mut BytesMut) {
         self.pdu_header.serialize(buf);
         self.originating_entity_id.serialize(buf);
         self.receiving_entity_id.serialize(buf);
         buf.put_u32(self.request_id as u32);
+        buf.put_u32(self.action_id as u32);
+        buf.put_u32(self.number_of_fixed_datum_records);
+        buf.put_u32(self.number_of_variable_datum_records);
+        buf.put_u64(self.fixed_datum_records);
+        buf.put_u64(self.variable_datum_records);
     }
 
     fn deserialize(mut buffer: BytesMut) -> Result<Self, DISError>
@@ -43,16 +59,32 @@ impl Pdu for CreateEntityPdu {
         Self: Sized,
     {
         let pdu_header = PduHeader::decode(&mut buffer);
-        if pdu_header.pdu_type == PduType::CreateEntity {
+        if pdu_header.pdu_type == PduType::ActionRequest {
             let originating_entity_id = EntityId::decode(&mut buffer);
             let receiving_entity_id = EntityId::decode(&mut buffer);
             let request_id = buffer.get_u32();
+            let action_id = buffer.get_u32();
+            let number_of_fixed_datum_records = buffer.get_u32();
+            let number_of_variable_datum_records = buffer.get_u32();
+            let mut fixed_datum_records: u64 = 0;
+            for _record in 0..number_of_fixed_datum_records as usize {
+                fixed_datum_records += buffer.get_u64();
+            }
+            let mut variable_datum_records: u64 = 0;
+            for _record in 0..number_of_variable_datum_records as usize {
+                variable_datum_records += buffer.get_u64();
+            }
 
-            return Ok(CreateEntityPdu {
+            return Ok(ActionRequestPdu {
                 pdu_header,
                 originating_entity_id,
                 receiving_entity_id,
                 request_id,
+                action_id,
+                number_of_fixed_datum_records,
+                number_of_variable_datum_records,
+                fixed_datum_records,
+                variable_datum_records,
             });
         } else {
             Err(DISError::InvalidDISHeader)
@@ -73,26 +105,42 @@ impl Pdu for CreateEntityPdu {
         let originating_entity_id = EntityId::decode(&mut buffer);
         let receiving_entity_id = EntityId::decode(&mut buffer);
         let request_id = buffer.get_u32();
+        let action_id = buffer.get_u32();
+        let number_of_fixed_datum_records = buffer.get_u32();
+        let number_of_variable_datum_records = buffer.get_u32();
+        let mut fixed_datum_records: u64 = 0;
+        for _record in 0..number_of_fixed_datum_records as usize {
+            fixed_datum_records += buffer.get_u64();
+        }
+        let mut variable_datum_records: u64 = 0;
+        for _record in 0..number_of_variable_datum_records as usize {
+            variable_datum_records += buffer.get_u64();
+        }
 
-        return Ok(CreateEntityPdu {
+        return Ok(ActionRequestPdu {
             pdu_header,
             originating_entity_id,
             receiving_entity_id,
             request_id,
+            action_id,
+            number_of_fixed_datum_records,
+            number_of_variable_datum_records,
+            fixed_datum_records,
+            variable_datum_records,
         });
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::CreateEntityPdu;
-    use crate::dis::common::pdu_header::{PduHeader, PduType, ProtocolFamily};
+    use super::ActionRequestPdu;
+    use crate::common::pdu_header::{PduHeader, PduType, ProtocolFamily};
 
     #[test]
     fn create_header() {
-        let action_request_pdu = CreateEntityPdu::default();
+        let action_request_pdu = ActionRequestPdu::default();
         let pdu_header = PduHeader::default(
-            PduType::CreateEntity,
+            PduType::ActionRequest,
             ProtocolFamily::SimulationManagement,
             448 / 8,
         );
