+++ src/simulation_management/data_query_pdu.rs
@@ -1,4 +1,4 @@
-use crate::dis::common::{
+use crate::common::{
     dis_error::DISError,
     entity_id::EntityId,
     pdu::Pdu,
@@ -8,26 +8,30 @@ use bytes::{Buf, BufMut, BytesMut};
 use std::any::Any;
 
 #[derive(Copy, Clone, Debug)]
-pub struct CommentPdu {
+pub struct DataQueryPdu {
     pub pdu_header: PduHeader,
     pub originating_entity_id: EntityId,
     pub receiving_entity_id: EntityId,
+    pub request_id: u32,
+    pub time_interval: u32,
     pub number_of_fixed_datum_records: u32,
     pub number_of_variable_datum_records: u32,
     pub fixed_datum_records: u64,
     pub variable_datum_records: u64,
 }
 
-impl CommentPdu {
+impl DataQueryPdu {
     pub fn default() -> Self {
-        CommentPdu {
+        DataQueryPdu {
             pdu_header: PduHeader::default(
-                PduType::Comment,
+                PduType::DataQuery,
                 ProtocolFamily::SimulationManagement,
                 56,
             ),
             originating_entity_id: EntityId::default(1),
             receiving_entity_id: EntityId::default(2),
+            request_id: 0,
+            time_interval: 0,
             number_of_fixed_datum_records: 0,
             number_of_variable_datum_records: 0,
             fixed_datum_records: 0,
@@ -36,11 +40,13 @@ impl CommentPdu {
     }
 }
 
-impl Pdu for CommentPdu {
+impl Pdu for DataQueryPdu {
     fn serialize(&self, buf: &mut BytesMut) {
         self.pdu_header.serialize(buf);
         self.originating_entity_id.serialize(buf);
         self.receiving_entity_id.serialize(buf);
+        buf.put_u32(self.request_id as u32);
+        buf.put_u32(self.time_interval as u32);
         buf.put_u32(self.number_of_fixed_datum_records);
         buf.put_u32(self.number_of_variable_datum_records);
         buf.put_u64(self.fixed_datum_records);
@@ -52,9 +58,11 @@ impl Pdu for CommentPdu {
         Self: Sized,
     {
         let pdu_header = PduHeader::decode(&mut buffer);
-        if pdu_header.pdu_type == PduType::Comment {
+        if pdu_header.pdu_type == PduType::DataQuery {
             let originating_entity_id = EntityId::decode(&mut buffer);
             let receiving_entity_id = EntityId::decode(&mut buffer);
+            let request_id = buffer.get_u32();
+            let time_interval = buffer.get_u32();
             let number_of_fixed_datum_records = buffer.get_u32();
             let number_of_variable_datum_records = buffer.get_u32();
             let mut fixed_datum_records: u64 = 0;
@@ -66,10 +74,12 @@ impl Pdu for CommentPdu {
                 variable_datum_records += buffer.get_u64();
             }
 
-            return Ok(CommentPdu {
+            return Ok(DataQueryPdu {
                 pdu_header,
                 originating_entity_id,
                 receiving_entity_id,
+                request_id,
+                time_interval,
                 number_of_fixed_datum_records,
                 number_of_variable_datum_records,
                 fixed_datum_records,
@@ -93,6 +103,8 @@ impl Pdu for CommentPdu {
     {
         let originating_entity_id = EntityId::decode(&mut buffer);
         let receiving_entity_id = EntityId::decode(&mut buffer);
+        let request_id = buffer.get_u32();
+        let time_interval = buffer.get_u32();
         let number_of_fixed_datum_records = buffer.get_u32();
         let number_of_variable_datum_records = buffer.get_u32();
         let mut fixed_datum_records: u64 = 0;
@@ -104,10 +116,12 @@ impl Pdu for CommentPdu {
             variable_datum_records += buffer.get_u64();
         }
 
-        return Ok(CommentPdu {
+        return Ok(DataQueryPdu {
             pdu_header,
             originating_entity_id,
             receiving_entity_id,
+            request_id,
+            time_interval,
             number_of_fixed_datum_records,
             number_of_variable_datum_records,
             fixed_datum_records,
@@ -118,14 +132,14 @@ impl Pdu for CommentPdu {
 
 #[cfg(test)]
 mod tests {
-    use super::CommentPdu;
-    use crate::dis::common::pdu_header::{PduHeader, PduType, ProtocolFamily};
+    use super::DataQueryPdu;
+    use crate::common::pdu_header::{PduHeader, PduType, ProtocolFamily};
 
     #[test]
     fn create_header() {
-        let action_request_pdu = CommentPdu::default();
+        let action_request_pdu = DataQueryPdu::default();
         let pdu_header = PduHeader::default(
-            PduType::Comment,
+            PduType::DataQuery,
             ProtocolFamily::SimulationManagement,
             448 / 8,
         );
