+++ src/simulation_management/stop_freeze_pdu.rs
@@ -1,7 +1,7 @@
 use bytes::{Buf, BufMut, BytesMut};
 use std::any::Any;
 
-use crate::dis::common::{
+use crate::common::{
     clock_time::ClockTime,
     dis_error::DISError,
     entity_id::EntityId,
@@ -174,7 +174,7 @@ impl FrozenBehavior {
 #[cfg(test)]
 mod tests {
     use super::StopFreezePdu;
-    use crate::dis::common::{
+    use crate::common::{
         pdu::Pdu,
         pdu_header::{PduHeader, PduType, ProtocolFamily},
     };
